# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API барои системаи тиҷоратии электронӣ",
        "description": "Санади API барои ҳама контроллерҳо",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/statuses": {
            "get": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Рӯйхати статусҳо",
                "operationId": "e96ea40c6f497546795df9ec612e1928",
                "parameters": [
                    {
                        "name": "for",
                        "in": "query",
                        "description": "Филтр (order ва ғайра)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Эҷоди статус",
                "operationId": "85739a0412fc324e86e6b45af20c7f0c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "for"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "object",
                                        "example": {
                                            "tj": "Нав",
                                            "ru": "Новый",
                                            "uz": "Yangi"
                                        }
                                    },
                                    "for": {
                                        "type": "string",
                                        "example": "order"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "new"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Сохта шуд"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/statuses/{id}": {
            "get": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Намоиши статус",
                "operationId": "ea554faff3dfcadd142b14010e8b3331",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Навсозии статус",
                "operationId": "7ba406b9e828f2ce3a0583d7c56583f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "object",
                                        "example": {
                                            "tj": "Нав",
                                            "ru": "Новый"
                                        }
                                    },
                                    "for": {
                                        "type": "string",
                                        "example": "order"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "new"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Навсозӣ шуд"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Нест кардани статус",
                "operationId": "4bcdd7d14d0707a62e62d2315c6149e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Нест шуд"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}